% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Say what animal makes what sound}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{A character string.}

\item{sound}{A character string.}
}
\value{
A character string saying what animal makes what sound.
}
\description{
A function which takes two character strings
and says what animal makes what sound.
Check out \link[base:paste]{paste0 here}!
}
